requirements.txt
selenium
pyautogui
pyperclip
pyyaml
opencv-python